move frl tkl to its own keyboard
