+++ keyboards/xelus/valor_frl_tkl/config.h
@@ -20,7 +20,7 @@
 
 /* USB Device descriptor parameter */
 #define VENDOR_ID       0x5845 // "XE"
-#define PRODUCT_ID      0x5653 // Valor "VR" + 1
+#define PRODUCT_ID      0x4654 // "FT"
 #define DEVICE_VER      0x0001
 #define MANUFACTURER    Xelus
 #define PRODUCT         Xelus Valor FRL TKL
