+++ keyboards/xelus/valor_frl_tkl/halconf.h
