+++ keyboards/xelus/valor_frl_tkl/readme.md
@@ -3,11 +3,11 @@
 An Alice-esque FRL TKL keyboard PCB.
 
 * Keyboard Maintainer: [Xelus22](https://github.com/Xelus22)
-* Hardware Supported: Valor
+* Hardware Supported: Valor FRL TKL
 * Hardware Availability: Custom keyboard group buys
 
 Make example for this keyboard (after setting up your build environment):
 
-    make xelus/valor/frl_tkl:default
+    make xelus/valor_frl_tkl:default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
