+++ keyboards/xelus/valor_frl_tkl/valor_frl_tkl.c
@@ -16,7 +16,7 @@
 
 // Nothing to see here, move along... ;-)
 
-#include "frl_tkl.h"
+#include "valor_frl_tkl.h"
 
 void board_init(void) {
   SYSCFG->CFGR1 |= SYSCFG_CFGR1_I2C1_DMA_RMP;
